adds grid and grid images selection

+++ resources/js/cms/modules/grid/Index.vue
@@ -1,64 +1,100 @@
 <template>
 <div v-if="isFetched">
-  
   <page-header>
     <h1>Bildraster Projekt</h1>
-    <a href="" class="btn-add has-icon">
+    <a href="" class="btn-add has-icon" @click.prevent="toggleGridSelector()">
       <plus-icon size="16"></plus-icon>
-      <span>Hinzufügen</span>
+      <span>Zeile</span>
     </a>
   </page-header>
 
-  <grid-layout-selector @select="addRow($event)" />
+  <grid-layout-selector 
+    @select="addRow($event)" 
+    v-if="hasGridSelector">
+  </grid-layout-selector>
+
+  <grid-image-selector
+   :images="$props.images"
+   @close="toggleImageSelector()"
+   @select="addImage($event)"
+   v-if="hasImageSelector">
+  </grid-image-selector>
 
   <grid-row 
     :key="grid.id" 
     v-for="grid in $props.grids" 
-    @deleteRow="deleteRow(grid.id)"
-    :class="`grid-layout grid-layout--${grid.layout}`"
-  >
+    :class="`grid-layout grid-layout--${grid.layout}`">
+    <a 
+      href="javascript:;" 
+      class="btn-delete has-icon"
+      @click.prevent="deleteRow(grid.id)">
+      <trash-2-icon size="16"></trash-2-icon>
+      <span>Löschen</span>
+    </a>
+    <div>
 
-    <template v-if="grid.layout == '1'">
-      <grid-row-item
-        v-for="item in grid.image_grid_items"
-        :key="item.id"
-        :image="item.image">
-      </grid-row-item>  
-    </template>
+      <template v-if="grid.layout == '1'">
+        <grid-row-item
+          v-for="item in grid.image_grid_items"
+          :key="item.id"
+          :item="item"
+          @showImages="showImageSelector($event)">
+        </grid-row-item>  
+      </template>
 
-    <template v-if="grid.layout == '1:1'">
-      <grid-row-item
-        v-for="item in grid.image_grid_items"
-        :key="item.id"
-        :image="item.image">
-      </grid-row-item>  
-    </template>
+      <template v-if="grid.layout == '1:1'">
+        <grid-row-item
+          v-for="item in grid.image_grid_items"
+          :key="item.id"
+          :item="item"
+          @showImages="showImageSelector($event)">
+        </grid-row-item>  
+      </template>
 
-    <template v-if="grid.layout == '1:2'">
-      <grid-row-item :image="grid.image_grid_items[0] ? grid.image_grid_items[0].image : null" />
-      <div class="grid-stack">
-        <grid-row-item :image="grid.image_grid_items[1] ? grid.image_grid_items[1].image : null" />
-        <grid-row-item :image="grid.image_grid_items[2] ? grid.image_grid_items[2].image : null" />
-      </div>
-    </template>
+      <template v-if="grid.layout == '1:2'">
+        <grid-row-item 
+          :item="grid.image_grid_items[0] ? grid.image_grid_items[0] : null"
+          @showImages="showImageSelector($event)">
+        </grid-row-item>
+        <div class="grid-stack">
+          <grid-row-item 
+            :item="grid.image_grid_items[1] ? grid.image_grid_items[1] : null"
+            @showImages="showImageSelector($event)">
+          </grid-row-item>
+          <grid-row-item 
+            :item="grid.image_grid_items[2] ? grid.image_grid_items[2] : null"
+            @showImages="showImageSelector($event)">
+          </grid-row-item>
+        </div>
+      </template>
 
-    <template v-if="grid.layout == '2:1'">
-      <div class="grid-stack">
-        <grid-row-item :image="grid.image_grid_items[0] ? grid.image_grid_items[0].image : null" />
-        <grid-row-item :image="grid.image_grid_items[1] ? grid.image_grid_items[1].image : null" />
-      </div>
-      <grid-row-item :image="grid.image_grid_items[2] ? grid.image_grid_items[2].image : null" />
-    </template>
+      <template v-if="grid.layout == '2:1'">
+        <div class="grid-stack">
+          <grid-row-item 
+            :item="grid.image_grid_items[0] ? grid.image_grid_items[0] : null"
+            @showImages="showImageSelector($event)">
+          </grid-row-item>
+          <grid-row-item 
+            :item="grid.image_grid_items[1] ? grid.image_grid_items[1] : null"
+            @showImages="showImageSelector($event)">
+          </grid-row-item>
+        </div>
+        <grid-row-item
+          :item="grid.image_grid_items[2] ? grid.image_grid_items[2] : null"
+          @showImages="showImageSelector($event)">
+        </grid-row-item>
+      </template>
 
-
+    </div>
   </grid-row>
 
 </div>
 </template>
 <script>
 import ErrorHandling from "@/mixins/ErrorHandling";
-import { PlusIcon } from 'vue-feather-icons';
+import { PlusIcon, Trash2Icon } from 'vue-feather-icons';
 import GridLayoutSelector from "@/modules/grid/components/GridLayoutSelector.vue";
+import GridImageSelector from "@/modules/grid/components/GridImageSelector.vue";
 import GridRow from "@/modules/grid/components/GridRow.vue";
 import GridRowItem from "@/modules/grid/components/GridRowItem.vue";
 import PageHeader from "@/components/ui/PageHeader.vue";
@@ -67,10 +103,12 @@ export default {
 
   components: {
     GridLayoutSelector,
+    GridImageSelector,
     GridRow,
     GridRowItem,
     PageHeader,
-    PlusIcon
+    PlusIcon,
+    Trash2Icon
   },
 
   mixins: [ErrorHandling],
@@ -78,6 +116,11 @@ export default {
   props: {
 
     grids: {
+      type: Array,
+      default: null,
+    },
+
+    images: {
       type: Array,
       default: null,
     },
@@ -99,37 +142,37 @@ export default {
       grid: [],
 
       currentRow: 0,
+      currentItemId: 0,
+      currentPos: 0,
 
       // Routes
       routes: {
         store: '/api/image-grid',
         order: '/api/image-grid/order',
-        delete: '/api/image-grid/row',
+        delete: '/api/image-grid',
+
+        storeItem: '/api/image-grid-item',
+
       },
 
       // States
       isLoading: false,
       isFetched: true,
+      hasGridSelector: false,
+      hasImageSelector: false,
 
       // Messages
       messages: {
         emptyData: 'Es sind noch keine Daten vorhanden...',
         confirm: 'Bitte löschen bestätigen!',
         updated: 'Daten aktualisiert',
-        stored: 'Daten gespeichert'
+        stored: 'Daten gespeichert',
+        deleted: 'Daten gelöscht',
       }
     }
   },
 
   methods: {
-
-    destroy(id) {
-      if (confirm(this.messages.confirm)) {
-        this.axios.delete(`${this.routes.delete}/${id}`).then(response => {
-          this.fetch();
-        });
-      }
-    },
 
     order(data) {
       let events = data.map(function(event, index) {
@@ -159,55 +202,55 @@ export default {
       this.axios.post(this.routes.store, grid).then(response => {
         this.$notify({ type: "success", text: this.messages.stored });
         this.isLoading = false;
+        this.$emit('addedRow');
       });
-
-      // // Get the last row to determine its layout and order
-      // const lastRow = this.grid[this.grid.length - 1];
-
-      // // Build next row
-      // const nextRow = {
-      //   layout: lastRow && lastRow.layout == '1-2' ? '2-1' : '1-2',
-      //   order: lastRow ? lastRow.order + 1 : 0
-      // };
-
-      // this.axios.post(this.routes.store, nextRow).then(response => {
-      //   this.fetch();
-      // });
     },
 
     deleteRow(id) {
-      this.axios.delete(`${this.routes.delete}/${id}`).then(response => {
-        const index = this.grid.findIndex(x => x.id == id);
-        this.grid.splice(index, 1);
+      if (confirm(this.messages.confirm)) {
+        this.isLoading = true;
+        this.axios.delete(`${this.routes.delete}/${id}`).then(response => {
+          this.$notify({ type: "success", text: this.messages.deleted });
+          this.isLoading = false;
+          this.$emit('deletedRow');
+        });
+      }
+    },
+
+    addImage(image) {
+      const item = {
+        id: this.currentItemId,
+        position: this.currentPos,
+        image_id: image
+      }
+      this.isLoading = true;
+      this.axios.post(this.routes.storeItem, item).then(response => {
+        this.$notify({ type: "success", text: this.messages.stored });
+        this.isLoading = false;
+        this.hideImageSelector();
+        this.$emit('addedRowItem');
       });
     },
 
-    resetItem(id) {
-      this.axios.get(`${this.routes.resetItem}/${id}`).then(response => {
-        this.fetch();
-      });
+    toggleGridSelector() {
+      this.hasGridSelector = this.hasGridSelector ? false : true;
     },
 
-    isEmpty(item) {
-      return false;
+    toggleImageSelector() {
+      this.hasImageSelector = this.hasImageSelector ? false : true;
+    },
+
+    showImageSelector(data) {
+      this.currentItemId = data.id;
+      this.currentPos = data.position;
+      this.hasImageSelector = true;
+    },
+
+    hideImageSelector() {
+      this.currentItemId = null;
+      this.currentPos = null;
+      this.hasImageSelector = false;
     }
   }
 }
 </script>
-<style scoped>
-.grid-buttons {
-  min-height: inherit;
-}
-
-.btn-primary {
-  max-width: 240px;
-  margin: 4px 0;
-}
-
-.btn-danger {
-  bottom: 4px;
-  max-width: 120px;
-  right: 4px;
-  position: absolute;
-}
-</style>
